feat: productList 관련 로직 별도 파일로 분리
